refactor: useDragDrop step 3 — cursor helper and calendar hover/click
